+++ src/signalling/room/command_handler.rs
@@ -1,7 +1,7 @@
 //! Implementation of the [`CommandHandler`] for the [`Room`] and related
 //! definitions.
 
-use std::collections::HashMap;
+use std::{collections::HashMap, convert::TryInto as _};
 
 use medea_client_api_proto as proto;
 use medea_client_api_proto::{
@@ -55,7 +55,10 @@ impl CommandHandler for Room {
 
         let from_member_id = from_peer.member_id();
         let to_member_id = to_peer.member_id();
-        let ice_servers = to_peer.ice_servers_list();
+        let ice_servers = to_peer
+            .ice_users()
+            .try_into()
+            .map_err(|_| RoomError::NoTurnCredentials(to_member_id.clone()))?;
 
         let event = if from_peer.is_known_to_remote() {
             Event::PeerUpdated {