+++ src/signalling/room/peer_events_handler.rs
@@ -1,5 +1,7 @@
 //! [`PeerConnectionStateEventsHandler`] implementation for [`Room`].
 
+use std::convert::TryInto as _;
+
 use actix::{Handler, Message, StreamHandler, WeakAddr};
 use chrono::{DateTime, Utc};
 use medea_client_api_proto::{
@@ -37,7 +39,14 @@ impl Room {
                 }
             };
 
-        let ice_servers = peer.ice_servers_list();
+        let ice_servers = if let Ok(ice_servers) = peer.ice_users().try_into() {
+            ice_servers
+        } else {
+            let member_id = peer.member_id().clone();
+            self.peers.add_peer(peer);
+            self.peers.add_peer(partner_peer);
+            return Err(RoomError::NoTurnCredentials(member_id));
+        };
 
         let peer = peer.start_as_offerer();
         let partner_peer = partner_peer.start_as_answerer();