+++ src/turn/ice_user.rs
@@ -4,12 +4,60 @@
 //! [TURN]: https://webrtcglossary.com/turn/
 //! [STUN]: https://webrtcglossary.com/stun/
 
+use std::convert::TryFrom;
+
 use derive_more::From;
 use medea_client_api_proto::IceServer;
 
 use crate::turn::static_service::StaticIceUser;
 
 use super::coturn::CoturnIceUser;
+
+/// Error which indicates that [`IceUsers`] is empty.
+#[derive(Debug)]
+pub struct EmptyIceServersListErr;
+
+/// List of [`IceUser`] created for some [`Peer`].
+///
+/// [`Peer`]: crate::media::peer::Peer
+#[derive(Debug)]
+pub struct IceUsers(Vec<IceUser>);
+
+impl IceUsers {
+    /// Returns new empty [`IceUsers`] list.
+    #[inline]
+    #[must_use]
+    pub fn new() -> Self {
+        Self(Vec::new())
+    }
+
+    /// Adds provided [`IceUser`]s to this [`IceUsers`] list.
+    #[inline]
+    pub fn add(&mut self, mut users: Vec<IceUser>) {
+        self.0.append(&mut users);
+    }
+}
+
+impl TryFrom<&IceUsers> for Vec<IceServer> {
+    type Error = EmptyIceServersListErr;
+
+    /// Returns vector of [`IceServer`]s if [`IceUsers`] list is not empty.
+    ///
+    /// Otherwise, returns [`EmptyIceServersListErr`].
+    fn try_from(value: &IceUsers) -> Result<Self, Self::Error> {
+        if value.0.is_empty() {
+            Err(EmptyIceServersListErr)
+        } else {
+            Ok(value.0.iter().flat_map(IceUser::servers_list).collect())
+        }
+    }
+}
+
+impl Default for IceUsers {
+    fn default() -> Self {
+        Self::new()
+    }
+}
 
 /// Credentials on Turn server.
 #[derive(Debug, From)]